first/requirements.txt
aiogram
pydantic-settings
openai
httpx
redis
sqlalchemy
asyncpg
alembic
amplitude-analytics